[Git page]Step 1: Added few git functions into another file and imported them. Status: Everything works!
